Modified several drawing algorithms to call OpenGL with floats as most GPU's don't have native support for doubles, initial implementation of cube drawing using connected triangle strips

+++ src/main/java/org/cytoscape/paperwing/internal/rendering/ResetSceneProcedure.java
@@ -30,8 +30,9 @@ public class ResetSceneProcedure implements ReadOnlyGraphicsProcedure {
 		Vector3 target = camera.getTarget();
 		Vector3 up = camera.getUp();
 
-		glu.gluLookAt(position.x(), position.y(), position.z(), target.x(),
-				target.y(), target.z(), up.x(), up.y(), up.z());
+		glu.gluLookAt((float) position.x(), (float) position.y(), (float) position.z(), 
+				(float) target.x(), (float) target.y(), (float) target.z(), 
+				(float) up.x(), (float) up.y(), (float) up.z());
 	}
 
 }
